test(backend): split investment tests into handler binding and service logic
Move the 9 binding cases (input validation, status mapping, partial
failure smoke) verbatim to api/investment_handler_test.go. Re-express
the rent-decline trend cases as direct service tests and add coverage
for AI summary attachment and the all-years-error path. Extend
stubSummarizer to return a configurable investment summary.
