+++ backend/internal/api/handler_test.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
@@ -440,61 +439,6 @@ func TestDomainValidate_Combinations(t *testing.T) {
 				t.Errorf("expected no error, got: %v", err)
 			}
 		})
-	}
-}
-
-func TestAnalyze_ValidInput(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	body, _ := json.Marshal(validBase)
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/analyze", bytes.NewReader(body))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-	var result domain.InvestmentResult
-	if err := json.NewDecoder(w.Body).Decode(&result); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if result.TotalInvestment <= 0 {
-		t.Error("expected totalInvestment > 0")
-	}
-}
-
-func TestAnalyze_InvalidJSON(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/analyze", bytes.NewBufferString("not-json"))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d", w.Code)
-	}
-}
-
-func TestAnalyze_ValidationError(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	invalid := validBase
-	invalid.LandPrice = -1
-	body, _ := json.Marshal(invalid)
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/analyze", bytes.NewReader(body))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d", w.Code)
-	}
-	var resp map[string]string
-	_ = json.NewDecoder(w.Body).Decode(&resp)
-	if resp["error"] == "" {
-		t.Error("expected error message in response")
 	}
 }
 
@@ -1126,295 +1070,6 @@ func TestUpstreamErrorNotExposed(t *testing.T) {
 				t.Errorf("response body %q must contain %q", body, tc.wantMessage)
 			}
 		})
-	}
-}
-
-// ---- MonteCarlo (/api/investment/simulate) ----
-
-func TestMonteCarlo_ValidInput(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	input := domain.MonteCarloInput{
-		Base:        validBase,
-		Simulations: 100,
-	}
-	body, _ := json.Marshal(input)
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/simulate", bytes.NewReader(body))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-	var result domain.MonteCarloResult
-	if err := json.NewDecoder(w.Body).Decode(&result); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if result.SimulationCount != 100 {
-		t.Errorf("expected simulationCount=100, got %d", result.SimulationCount)
-	}
-	if result.SuccessRate < 0 || result.SuccessRate > 1 {
-		t.Errorf("successRate out of range: %f", result.SuccessRate)
-	}
-}
-
-func TestMonteCarlo_InvalidJSON(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/simulate", bytes.NewBufferString("not-json"))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d", w.Code)
-	}
-	var resp map[string]string
-	_ = json.NewDecoder(w.Body).Decode(&resp)
-	if resp["error"] == "" {
-		t.Error("expected error message in response body")
-	}
-}
-
-func TestMonteCarlo_ValidationError(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-
-	invalid := validBase
-	invalid.LandPrice = -1 // invalid
-	input := domain.MonteCarloInput{
-		Base:        invalid,
-		Simulations: 100,
-	}
-	body, _ := json.Marshal(input)
-	req := httptest.NewRequest(http.MethodPost, "/api/investment/simulate", bytes.NewReader(body))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d", w.Code)
-	}
-	var resp map[string]string
-	_ = json.NewDecoder(w.Body).Decode(&resp)
-	if resp["error"] == "" {
-		t.Error("expected error message in response")
-	}
-}
-
-// ---- GetRentDeclineHint (/api/investment/rent-decline-hint) ----
-
-func TestGetRentDeclineHint_MissingArea(t *testing.T) {
-	r := newTestRouter(&mockMLITClient{}, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d: %s", w.Code, w.Body.String())
-	}
-	var resp map[string]string
-	_ = json.NewDecoder(w.Body).Decode(&resp)
-	if resp["error"] == "" {
-		t.Error("expected error message in response body")
-	}
-}
-
-func TestGetRentDeclineHint_AllYearsAPIError(t *testing.T) {
-	client := &mockMLITClient{
-		appraisalFunc: func(_ context.Context, _, _ string, _ int, _ string) ([]domain.LandAppraisalItem, error) {
-			return nil, errors.New("upstream error")
-		},
-	}
-	r := newTestRouter(client, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint?area=13", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadGateway {
-		t.Fatalf("expected 502, got %d: %s", w.Code, w.Body.String())
-	}
-}
-
-func TestGetRentDeclineHint_PartialYearError(t *testing.T) {
-	// 2022, 2023 はエラー（5年中2年）、2024〜2026 は有効な下落データを返す
-	client := &mockMLITClient{
-		appraisalFunc: func(_ context.Context, _, _ string, year int, _ string) ([]domain.LandAppraisalItem, error) {
-			switch year {
-			case 2022, 2023:
-				return nil, errors.New("upstream error")
-			case 2024:
-				return []domain.LandAppraisalItem{
-					{Year: 2024, PricePerSqm: 180000, ChangeRate: -0.04},
-					{Year: 2024, PricePerSqm: 185000, ChangeRate: -0.04},
-					{Year: 2024, PricePerSqm: 175000, ChangeRate: -0.04},
-					{Year: 2024, PricePerSqm: 182000, ChangeRate: -0.04},
-					{Year: 2024, PricePerSqm: 178000, ChangeRate: -0.04},
-				}, nil
-			case 2025:
-				return []domain.LandAppraisalItem{
-					{Year: 2025, PricePerSqm: 170000, ChangeRate: -0.03},
-					{Year: 2025, PricePerSqm: 175000, ChangeRate: -0.03},
-					{Year: 2025, PricePerSqm: 168000, ChangeRate: -0.03},
-					{Year: 2025, PricePerSqm: 172000, ChangeRate: -0.03},
-					{Year: 2025, PricePerSqm: 165000, ChangeRate: -0.03},
-				}, nil
-			case 2026:
-				return []domain.LandAppraisalItem{
-					{Year: 2026, PricePerSqm: 162000, ChangeRate: -0.02},
-					{Year: 2026, PricePerSqm: 165000, ChangeRate: -0.02},
-					{Year: 2026, PricePerSqm: 160000, ChangeRate: -0.02},
-					{Year: 2026, PricePerSqm: 163000, ChangeRate: -0.02},
-					{Year: 2026, PricePerSqm: 158000, ChangeRate: -0.02},
-				}, nil
-			default:
-				return []domain.LandAppraisalItem{}, nil
-			}
-		},
-	}
-	r := newTestRouter(client, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint?area=13", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	// 一部年エラーでも有効データがあれば200を返す
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200 for partial year error, got %d: %s", w.Code, w.Body.String())
-	}
-	var hint domain.RentDeclineHint
-	if err := json.NewDecoder(w.Body).Decode(&hint); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if hint.Basis != "land_appraisal" {
-		t.Errorf("expected basis=land_appraisal, got %q", hint.Basis)
-	}
-}
-
-func TestGetRentDeclineHint_DeclineTrend(t *testing.T) {
-	// 2022〜2024の3年分、各年5件ずつ地価下落傾向のデータを返す
-	client := &mockMLITClient{
-		appraisalFunc: func(_ context.Context, _, _ string, year int, _ string) ([]domain.LandAppraisalItem, error) {
-			switch year {
-			case 2022:
-				return []domain.LandAppraisalItem{
-					{Year: 2022, PricePerSqm: 200000, ChangeRate: -0.02},
-					{Year: 2022, PricePerSqm: 210000, ChangeRate: -0.02},
-					{Year: 2022, PricePerSqm: 190000, ChangeRate: -0.02},
-					{Year: 2022, PricePerSqm: 205000, ChangeRate: -0.02},
-					{Year: 2022, PricePerSqm: 195000, ChangeRate: -0.02},
-				}, nil
-			case 2024:
-				return []domain.LandAppraisalItem{
-					{Year: 2024, PricePerSqm: 180000, ChangeRate: -0.05},
-					{Year: 2024, PricePerSqm: 185000, ChangeRate: -0.05},
-					{Year: 2024, PricePerSqm: 175000, ChangeRate: -0.05},
-					{Year: 2024, PricePerSqm: 182000, ChangeRate: -0.05},
-					{Year: 2024, PricePerSqm: 178000, ChangeRate: -0.05},
-				}, nil
-			default:
-				return []domain.LandAppraisalItem{}, nil
-			}
-		},
-	}
-	r := newTestRouter(client, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint?area=13", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-	var hint domain.RentDeclineHint
-	if err := json.NewDecoder(w.Body).Decode(&hint); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if hint.Basis != "land_appraisal" {
-		t.Errorf("expected basis=land_appraisal, got %q", hint.Basis)
-	}
-	if hint.HintRate <= 0 {
-		t.Errorf("expected hintRate > 0 for declining land prices, got %f", hint.HintRate)
-	}
-	if hint.FallbackUsed {
-		t.Error("expected fallbackUsed=false for land_appraisal basis")
-	}
-}
-
-func TestGetRentDeclineHint_RisingTrend(t *testing.T) {
-	// 地価上昇傾向 → fallback を返す
-	client := &mockMLITClient{
-		appraisalFunc: func(_ context.Context, _, _ string, year int, _ string) ([]domain.LandAppraisalItem, error) {
-			switch year {
-			case 2022:
-				return []domain.LandAppraisalItem{
-					{Year: 2022, PricePerSqm: 100000, ChangeRate: 0.05},
-					{Year: 2022, PricePerSqm: 100000, ChangeRate: 0.05},
-					{Year: 2022, PricePerSqm: 100000, ChangeRate: 0.05},
-					{Year: 2022, PricePerSqm: 100000, ChangeRate: 0.05},
-					{Year: 2022, PricePerSqm: 100000, ChangeRate: 0.05},
-				}, nil
-			case 2024:
-				return []domain.LandAppraisalItem{
-					{Year: 2024, PricePerSqm: 120000, ChangeRate: 0.05},
-					{Year: 2024, PricePerSqm: 120000, ChangeRate: 0.05},
-					{Year: 2024, PricePerSqm: 120000, ChangeRate: 0.05},
-					{Year: 2024, PricePerSqm: 120000, ChangeRate: 0.05},
-					{Year: 2024, PricePerSqm: 120000, ChangeRate: 0.05},
-				}, nil
-			default:
-				return []domain.LandAppraisalItem{}, nil
-			}
-		},
-	}
-	r := newTestRouter(client, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint?area=13", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-	var hint domain.RentDeclineHint
-	if err := json.NewDecoder(w.Body).Decode(&hint); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if hint.Basis != "fallback" {
-		t.Errorf("expected basis=fallback for rising prices, got %q", hint.Basis)
-	}
-	if !hint.FallbackUsed {
-		t.Error("expected fallbackUsed=true for rising prices")
-	}
-}
-
-func TestGetRentDeclineHint_InsufficientData(t *testing.T) {
-	// データが5件未満 → fallback
-	client := &mockMLITClient{
-		appraisalFunc: func(_ context.Context, _, _ string, year int, _ string) ([]domain.LandAppraisalItem, error) {
-			if year == 2024 {
-				// 4件だけ返す（minDataPointsForHint=5未満）
-				return []domain.LandAppraisalItem{
-					{Year: 2024, PricePerSqm: 150000, ChangeRate: -0.01},
-					{Year: 2024, PricePerSqm: 160000, ChangeRate: -0.01},
-					{Year: 2024, PricePerSqm: 155000, ChangeRate: -0.01},
-					{Year: 2024, PricePerSqm: 145000, ChangeRate: -0.01},
-				}, nil
-			}
-			return []domain.LandAppraisalItem{}, nil
-		},
-	}
-	r := newTestRouter(client, nil)
-	req := httptest.NewRequest(http.MethodGet, "/api/investment/rent-decline-hint?area=13", nil)
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-	var hint domain.RentDeclineHint
-	if err := json.NewDecoder(w.Body).Decode(&hint); err != nil {
-		t.Fatalf("failed to decode response: %v", err)
-	}
-	if hint.Basis != "fallback" {
-		t.Errorf("expected basis=fallback for insufficient data, got %q", hint.Basis)
 	}
 }
